[CP Optimizer] Add redundant cumulative constraints
