+++ pyjobshop/solvers/cpoptimizer/Constraints.py
@@ -244,6 +244,35 @@ class Constraints:
 
             model.add(expr1 <= expr2)
 
+    def _redundant_cumulative_constraints(self):
+        """
+        Adds redundant cumulative constraints for connected components of
+        resources.
+        """
+        model, data = self._model, self._data
+
+        for component in utils.connected_resource_components(data):
+            if len(component) < 1 or len(component) == data.num_resources:
+                continue
+
+            intervals = []
+            for task_idx, task_var in enumerate(self._task_vars):
+                task_resources = {
+                    res
+                    for mode in data.task2modes(task_idx)
+                    for res in data.modes[mode].resources
+                }
+                if component & task_resources:
+                    # Collect intervals for tasks that are assigned to
+                    # resource in the component.
+                    intervals.append(task_var)
+
+            if len(intervals) <= 1 or len(intervals) == data.num_tasks:
+                continue
+
+            pulses = [cpo.pulse(interval, 1) for interval in intervals]
+            model.add(sum(pulses) <= len(component))
+
     def add_constraints(self):
         """
         Adds all the constraints to the CP model.
@@ -258,3 +287,4 @@ class Constraints:
         self._identical_and_different_resource_constraints()
         self._consecutive_constraints()
         self._mode_dependencies()
+        self._redundant_cumulative_constraints()
